Fix places where the "float" type is used
This is almost never desired, because it means that only 24 bits
of precision are kept.  Usually this was done by accident, when
thinking about the HAL type hal_float, which is actually a "double
precision" type too.

+++ src/hal/drivers/mesa-hostmot2/sserial.c
@@ -1776,7 +1776,7 @@ void hm2_sserial_write_pins(hostmot2_t *hm2, hm2_sserial_instance_t *inst){
     int b, p, r;
     int bitcount;
     rtapi_u64 buff;
-    float val;
+    double val;
 
     // the side effect of reporting this error will suffice
     (void)hm2_sserial_check_remote_errors(hm2, inst);
